refactor: SubHeader 구조 및 버튼 레이아웃 개선
